인생건 최적화

+++ Assets/Scripts/Piece/SkillPieces/Nature/Skill_N_Swamp.cs
@@ -65,7 +65,7 @@ public class Skill_N_Swamp : SkillPiece
             .Play();
 
             }, BezierType.Quadratic, isRotate: true, playSpeed: 1.5f);
-            yield return new WaitForSeconds(time / (float)rand);
+            yield return new WaitForSeconds(time / rand);
         }
     }
 }
